LANGCHAIN-OLLAMA/requirements.txt
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chroma
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub